requirements.txt
# IMS Agent — Python dependencies
# Install with: pip install -r requirements.txt

# Anthropic SDK
anthropic>=0.40.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Environment variable loading
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0

# Type checking (dev)
mypy>=1.10.0
